requirements.txt
aiotdlib==0.22.0
python-dotenv==1.0.1
SQLAlchemy==2.0.29
psycopg2-binary==2.9.9
pydantic==1.10.15